requirements.txt
Flask>=2.1
flask-cors
google-genai>=1.0.0